RBigObject

+++ RBigObject/tests/test-all.R
@@ -1,24 +1,25 @@
 library("testthat")
 library("RBigObject")
-library("digest")
 library("magrittr")
 
-bigobject_connection(Sys.getenv("BIGOBJECT_IP"), Sys.getenv("BIGOBJECT_PORT"))
-bigobject_verbose(FALSE)
-
-{
-# bootstrap
-  invisible(bigobject_sql("DROP TABLE hashtest"))
-  invisible(bigobject_sql("DROP TABLE iristest"))
-}
-
-test_check("RBigObject")
-
-{
-  # clean up
-  invisible(bigobject_sql("DROP TABLE hashtest"))
-  invisible(bigobject_sql("DROP TABLE iristest"))
-  bigobject_sql("GC ALL")
+if (Sys.getenv("BIGOBJECT_IP") != "") {
+  bigobject_verbose(FALSE)
+  bigobject_connection(Sys.getenv("BIGOBJECT_IP"), Sys.getenv("BIGOBJECT_PORT"))
+  
+  {
+    # bootstrap
+    invisible(bigobject_sql("DROP TABLE hashtest"))
+    invisible(bigobject_sql("DROP TABLE iristest"))
+  }
+  
+  test_check("RBigObject")
+  
+  {
+    # clean up
+    invisible(bigobject_sql("DROP TABLE hashtest"))
+    invisible(bigobject_sql("DROP TABLE iristest"))
+    invisible(bigobject_sql("GC ALL"))
+  }
 }
